Move all three directories that do Context data stuff into src/contexts
