+++ src/components/DAOs/DAO/Proposals/CastVote.tsx
@@ -1,6 +1,6 @@
-import { ProposalData } from "../../../../daoData/useProposals";
+import { ProposalData } from "../../../../contexts/daoData/useProposals";
 import { useState, useEffect } from "react";
-import useCastVote from "../../../../daoData/useCastVote";
+import useCastVote from "../../../../contexts/daoData/useCastVote";
 import { PrimaryButton, SecondaryButton } from "../../../ui/forms/Button";
 import Check from "../../../ui/svg/Check";
 import ContentBanner from "../../../ui/ContentBanner";
