+++ src/contexts/daoData/useCastVote.ts
@@ -2,7 +2,7 @@ import { useCallback } from 'react'
 import { useTransaction } from '../web3Data/transactions';
 import { useWeb3 } from '../web3Data';
 import { BigNumber } from 'ethers';
-import { GovernorModule, GovernorModule__factory } from '../typechain-types';
+import { GovernorModule, GovernorModule__factory } from '../../typechain-types';
 import { useDAOData } from './index';
 
 const useCastVote = ({
