+++ src/contexts/daoData/useGovernorModuleContract.ts
@@ -1,5 +1,5 @@
 import { useState, useEffect } from "react";
-import { GovernorModule, GovernorModule__factory } from "../typechain-types";
+import { GovernorModule, GovernorModule__factory } from "../../typechain-types";
 import { useWeb3 } from "../web3Data";
 
 const useGovernorModuleContract = (moduleAddresses: string[] | undefined) => {
