+++ src/contexts/daoData/useProposals.ts
@@ -1,5 +1,5 @@
 import { useState, useEffect } from "react";
-import { GovernorModule } from "../typechain-types";
+import { GovernorModule } from "../../typechain-types";
 import { useWeb3 } from "../web3Data";
 import { BigNumber, providers } from "ethers";
