+++ src/contexts/daoData/useTimelockModuleContract.ts
@@ -1,5 +1,5 @@
 import { useState, useEffect } from "react";
-import { TimelockUpgradeable, TimelockUpgradeable__factory } from "../typechain-types";
+import { TimelockUpgradeable, TimelockUpgradeable__factory } from "../../typechain-types";
 import { useWeb3 } from "../web3Data";
 
 const useTimelockModuleContract = (moduleAddresses: string[] | undefined) => {
