+++ src/contexts/daoData/useTokenContract.ts
@@ -1,5 +1,5 @@
 import { useState, useEffect } from 'react';
-import { VotesTokenWithSupply, VotesTokenWithSupply__factory, GovernorModule } from '../typechain-types';
+import { VotesTokenWithSupply, VotesTokenWithSupply__factory, GovernorModule } from '../../typechain-types';
 import { useWeb3 } from '../web3Data';
 
 const useTokenContract = (governorModule: GovernorModule | undefined) => {
